Fix up some sloppy hacks from before.

+++ vr4300/fault.c
@@ -137,7 +137,7 @@ void VR4300_DCB(struct vr4300 *vr4300) {
         unsigned rshiftamt = (4 - request->size) << 3;
         unsigned lshiftamt = (paddr & 0x3) << 3;
 
-        request->data = ((int64_t) (word << lshiftamt)) >> rshiftamt;
+        request->data = (int32_t) (word << lshiftamt) >> rshiftamt;
       }
 
       else {
